django_projects/blog_api/requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
pillow
dj_rest_auth